+++ app/policies/users_policy.rb
@@ -1,8 +1,0 @@
-class UsersPolicy < ApplicationPolicy
-
-  primary_resource :user
-
-  def show?
-    current_user == user
-  end
-end