+++ engines/plos_billing/spec/models/billing/billing_task_spec.rb
@@ -8,14 +8,12 @@ module PlosBilling
         completed: true,
         paper: paper,
         phase: paper.phases.first,
-        title: "Billing",
-        old_role: "author"
+        title: "Billing"
       )
     end
 
     describe '.restore_defaults' do
       it_behaves_like '<Task class>.restore_defaults update title to the default'
-      it_behaves_like '<Task class>.restore_defaults update old_role to the default'
     end
 
     describe '.create' do