+++ engines/tahi_standard_tasks/app/models/tahi_standard_tasks/taxon_task.rb
@@ -1,7 +1,7 @@
 module TahiStandardTasks
   class TaxonTask < ::Task
     include MetadataTask
-    DEFAULT_TITLE = 'New Taxon'
-    DEFAULT_ROLE = 'author'
+    DEFAULT_TITLE = 'New Taxon'.freeze
+    DEFAULT_ROLE_HINT = 'author'.freeze
   end
 end